Add support for input and output trancoder config
In order to support different transcoding configuration, it is now
possible to output the default configure, modify it and pass a new
configuration that will override the default one.

This supports both yaml (for people) and json (for machines).

Add 4 new arguments:
--output-yml-config: output yml default configuration to stdout
--output-json-config: output json default configuration to stdout
--input-yml-config: input yml configuration from specified path
--input-json-config: input json configuration from specified path

Add 2 basic test cases for outputing and inputing default configuration.
Refactor the main script to be able to do it.